Get extended attributes for user for apps
Get extended attributes for user from the apps. A
new method which emits events to get the attributes
for user, which is listened by apps to provide the
attributes for the specific user. Use case for example:
When apps are not whitelisted for guest user
they should not be visible in the personal settings.

Signed-off-by: Sujith H <sharidasan@owncloud.com>